+++ Code/MCR_2Code.lua
@@ -5,12 +5,13 @@
 -- Author @SkiRich
 -- This mod is subject to the terms and conditions outlined in the LICENSE file included in this mod.
 -- Created Oct 14th, 2018
--- Updated Dec 3rd, 2018
+-- Updated Dec 11th, 2018
 
 local lf_debug   = false  -- used only for certain ex() instance
 local lf_print   = false  -- Setup debug printing in local file
                           -- Use if lf_print then print("something") end
 
+local ModConfigWarnThread = false -- var to keep track of warning thread
 
 ModConfig = {}    -- base class for modconfig
 ModConfig.StringIdBase = 76827146
@@ -51,7 +52,8 @@ end -- if type
 function ModConfig:RegisterMod(mod_id, mod_name, mod_desc, save_loc)
     mod_name = mod_name or mod_id
     mod_desc = mod_desc or ""
-    save_loc = save_loc or "centralstorage"
+    save_loc = save_loc or "centralstorage" -- set default to central storage
+    if (save_loc ~= "centralstorage") and (save_loc ~= "localstorage") then save_loc = "centralstorage" end -- only allow these two settings
     if not self.registry then self.registry = {} end
     self.registry[mod_id] = {name = mod_name, desc = mod_desc, save = save_loc}
 end -- ModConfig:RegisterMod
@@ -195,20 +197,45 @@ function ModConfig:GetDefault(mod_id, option_id)
     return mod_options and mod_options[option_id] and mod_options[option_id].default
 end -- ModConfig:GetDefault
 
+-- reset all mod config options back to default
 ----------------------------------- ModConfig:ResetAllToDefault ----------------------------------------
 function ModConfig:ResetAllToDefaults()
-	local registry = self.registry
-	ModConfig.data = {} -- blank out the data
-	for mod_id in pairs(registry) do
-		for option_id in pairs(registry[mod_id].options) do
-			local defaultvalue = ModConfig:GetDefault(mod_id, option_id)
-			if type(defaultvalue) ~= "nil" then
-				if lf_print then print("Resetting: ", mod_id, " ", option_id, " to ", tostring(defaultvalue)) end
-				ModConfig:Set(mod_id, option_id, defaultvalue)
-			end
-		end -- option_id
-  end -- for mod_id
-  ModConfig:Save()
+	-- use thread to slow up process slightly in case of File I/O and process pre-emptions
+	CreateRealTimeThread(function(self)
+		local reset = true
+	  local registry = self.registry
+	  self.data = {} -- blank out the data
+	  for mod_id in pairs(registry) do
+	  	for option_id in pairs(registry[mod_id].options) do
+	  		local defaultvalue = self:GetDefault(mod_id, option_id)
+	  		local optiontype = registry[mod_id].options[option_id].type
+	  		if (type(defaultvalue) ~= "nil") then
+	  			-- if there is a default, then use that
+	  			-- type 'enum' will use default or if not defined, the first item in the enum table
+	  			if lf_print then print("Resetting Default: ", mod_id, " ", option_id, " to ", tostring(defaultvalue)) end
+	  			if not self.data[mod_id] then self.data[mod_id] = {} end
+	  			-- since the value is default, we dont need to add the option to the the data
+	  			self.data[mod_id][option_id] = nil
+	  			self:Save()
+	  			-- we still need to broadcast the options default value since things change based on value
+	  			Msg("ModConfigChanged", mod_id, option_id, defaultvalue, nil, "ModConfigReset")
+	  			Sleep(10)
+	  		elseif (type(defaultvalue) == "nil") and (optiontype ~= "note") then
+	  			-- if no default is set, use the current option (used for any missing defaults)
+	  			-- exclude 'note' type since thats not a settable option
+	  			local currentvalue = self:Get(mod_id, option_id)
+	  			if lf_print then print("Resetting NIL Default: ", mod_id, " ", option_id, " to ", tostring(currentvalue)) end
+	  			if not self.data[mod_id] then self.data[mod_id] = {} end
+	  			self.data[mod_id][option_id] = currentvalue
+	  			self:Save()
+	  			Msg("ModConfigChanged", mod_id, option_id, currentvalue, nil, "ModConfigReset")
+	  			Sleep(10)
+	  		end -- if defaultvalue
+	  	end -- for option_id
+    end -- for mod_id
+    ModLog("MCR - Reset Mod Config Data Complete")
+    self:OpenDialog()
+  end, self) -- CreateRealTimeThread
 end -- ModConfig:ResetAllToDefault()
 
 
@@ -318,8 +345,12 @@ function ModConfig:CalcDataSpace()
 	end -- for regMod
 
   local central_save_data = Compress(ValueToLuaCode(save_data))
-  local dataspace = 100* string.len(central_save_data) / const.MaxModDataSize
-  return dataspace
+  if lf_print then print("central_save_data len: ", string.len(central_save_data), " Max : ", const.MaxModDataSize) end
+
+  local overdatalimit = string.len(central_save_data) > const.MaxModDataSize
+  local dataspace = ((100 * string.len(central_save_data)) + 0.0) / const.MaxModDataSize
+  dataspace = string.format("%.2f", dataspace)
+  return dataspace, overdatalimit
 end -- ModConfig:CalcDataSpace()
 
 
@@ -352,6 +383,28 @@ function ModConfig:SaveSettingsFile()
   end -- if not err
 end -- ModConfig:SaveSettingsFile()
 
+----------------------------------- ModConfigWarnOverLimit() ------------------------------------------
+function ModConfigWarnOverLimit()
+	if not IsValidThread(ModConfigWarnThread) then
+    ModConfigWarnThread = CreateRealTimeThread(function()
+        local params = {
+            title = T{"Mod Config Reborn Warning"},
+             text = T{"Maximum persistent storage limit has been reached. There are too many options in Mod Config and/or too many options have been changed.<newline>"..
+             	        "A corrupt Mod Config database can also cause this.<newline>"..
+             	        "You can try to reset Mod Config Reborn to fix the issue, or contact the authors of the mods that use Mod Config Reborn.<newline>"..
+             	        "Changes in Mod Config Reborn are not being saved"},
+            choice1 = T{"OK"},
+            image = "UI/Messages/death.tga",
+            start_minimized = false,
+        } -- params
+        local choice = WaitPopupNotification(false, params)
+        if choice == 1 then
+          --nothing
+        end -- if statement
+        Sleep(60 * 1000) -- run thread for 1 minute so we dont spam the user.
+    end ) -- CreateRealTimeThread
+  end -- if not IsValidThread
+end -- function end
 
 ----------------------------------- ModConfig:Save -----------------------------------------------------
 -- Save all of the current settings to mcr modpersistant data.
@@ -362,18 +415,24 @@ function ModConfig:Save()
 	local registry = self.registry or {}
 
 	for regMod, regModOpt in pairs(registry) do
-		if ((type(regModOpt.save) == nil) or (regModOpt.save == "centralstorage")) and mod_data[regMod] then
+		if ((type(regModOpt.save) == "nil") or (regModOpt.save == "centralstorage")) and mod_data[regMod] then
 			save_data[regMod] = mod_data[regMod]
 		end -- if mod_data
 	end -- for regMod
 
+  local dataspace, overdatalimit = self:CalcDataSpace()
   local central_save_data = Compress(ValueToLuaCode(save_data))
   local interface = GetInGameInterface()
   if interface and interface.idModConfigDlg  and interface.idModConfigDlg:IsVisible() then
-      ModConfig.space_label:SetText(T{ModConfig.StringIdBase + 7, "Storage space in use: <used>%", used = ModConfig:CalcDataSpace() })
+      ModConfig.space_label:SetText(T{ModConfig.StringIdBase + 7, "Storage space in use: <used>%", used = dataspace })
   end -- if interface
-  WriteModPersistentData(central_save_data)
-  self:SaveSettingsFile()
+
+  -- check space in persistable memory and warn if over limit
+  if overdatalimit then ModConfigWarnOverLimit() end
+
+  -- only save data if datalimit is less then max
+  if not overdatalimit then WriteModPersistentData(central_save_data) end
+  self:SaveSettingsFile() -- we can always save to file
 end -- ModConfig:Save